Updating readme and examples, making rodio and realtime features

+++ ardftsrc/examples/rodio_adapter.rs
@@ -29,8 +29,8 @@ fn main() -> Result<(), Box<dyn Error>> {
         .with_channels(1)
         .with_output_rate(OUTPUT_SAMPLE_RATE_HZ)
         .with_input_rate(INPUT_SAMPLE_RATE_HZ);
-    let streaming_resampler = RealtimeResampler::<f32>::new(config, StreamingConfig::default());
-    let resampled_tone: RodioResampler<_, f32> = RodioResampler::new(tone, streaming_resampler);
+    let streaming_resampler = RealtimeResampler::<f64>::new(config, StreamingConfig::default());
+    let resampled_tone: RodioResampler<_, f64> = RodioResampler::new(tone, streaming_resampler);
 
     mixer.add(resampled_tone);
     thread::sleep(Duration::from_secs(DURATION_SECS));
