+++ ardftsrc/examples/rodio_adapter_with_spans.rs
@@ -42,8 +42,8 @@ fn main() -> Result<(), Box<dyn Error>> {
         channels: 1,
         ..Config::default()
     };
-    let streaming_resampler = RealtimeResampler::<f32>::new(config, StreamingConfig::default());
-    let resampled_tone: RodioResampler<_, f32> = RodioResampler::new(tone, streaming_resampler);
+    let streaming_resampler = RealtimeResampler::<f64>::new(config, StreamingConfig::default());
+    let resampled_tone: RodioResampler<_, f64> = RodioResampler::new(tone, streaming_resampler);
 
     mixer.add(resampled_tone);
     thread::sleep(Duration::from_secs(SPAN_DURATION_SECS * 2));
